requirements.txt
hydra-core
pytorch-lightning
wandb
torch
torchvision
torchmetrics
